added correct error messages

+++ src/contexts/video-action-context/video-action-utils/like-utils.js
@@ -1,11 +1,12 @@
 import axios from "axios";
-import { Toast } from "../../../components/index"
+import { Toast } from "../../../components/index";
 
-const isLiked = (id,data) => {
-  return data?.find(item=>item._id === id)
-}
+const isLiked = (id, data) => {
+  return data?.find(item => item._id === id);
+};
 
-const getLikedVideos = async (dispatchAction,navigate) => {
+const getLikedVideos = async (isLoggedIn, dispatchAction, navigate) => {
+  if (isLoggedIn) {
     try {
       const response = await axios.get("/api/user/likes", {
         headers: {
@@ -13,24 +14,37 @@ const getLikedVideos = async (dispatchAction,navigate) => {
         },
       });
       if (response.status === 200) {
-        dispatchAction({type:"UPDATE_LIKED_VIDEOS", payload:response.data.likes})
+        dispatchAction({
+          type: "UPDATE_LIKED_VIDEOS",
+          payload: response.data.likes,
+        });
       } else {
-        navigate("/signup");
+        Toast({
+          message: "Some error occured, please try again later",
+          type: "error",
+        });
       }
     } catch (err) {
-      navigate("/signup")
       Toast({
-        message: "Please login to continue.",
-        type: "warning",
+        message: "Some error occured, please try again later",
+        type: "error",
       });
     }
-  };
+  } else {
+    navigate("/signup");
+    Toast({
+      message: "Please login to continue.",
+      type: "warning",
+    });
+  }
+};
 
-const addToLiked = async (video,dispatchliked,navigate) => {
+const addToLiked = async (isLoggedIn, video, dispatchliked, navigate) => {
+  if (isLoggedIn) {
     try {
       const response = await axios.post(
         "/api/user/likes",
-        { video  },
+        { video },
         {
           headers: {
             authorization: localStorage.getItem("userToken"),
@@ -38,29 +52,37 @@ const addToLiked = async (video,dispatchliked,navigate) => {
         }
       );
       if (response.status === 201) {
-        dispatchliked({ type: "UPDATE_LIKED_VIDEOS", payload: response.data.likes });
+        dispatchliked({
+          type: "UPDATE_LIKED_VIDEOS",
+          payload: response.data.likes,
+        });
         Toast({
-        message: "Video added to your liked videos.",
-        type: "success",
-      });
-      }
-      else{
-          navigate("/signup")
-          Toast({
-        message: "Please login to continue.",
-        type: "warning",
-      });
+          message: "Video added to your liked videos.",
+          type: "success",
+        });
+      } else {
+        Toast({
+          message: "Some error occured, please try again later",
+          type: "error",
+        });
       }
     } catch (err) {
-     navigate("/signup")
-     Toast({
-        message: "Please login to continue.",
-        type: "warning",
+      Toast({
+        message: "Some error occured, please try again later",
+        type: "error",
       });
     }
-  };
+  } else {
+    navigate("/signup");
+    Toast({
+      message: "Please login to continue.",
+      type: "warning",
+    });
+  }
+};
 
-const removeFromliked = async (id,dispatchAction,navigate) => {
+const removeFromliked = async (isLoggedIn, id, dispatchAction, navigate) => {
+  if (isLoggedIn) {
     const path = `/api/user/likes/${id}`;
     try {
       const response = await axios.delete(path, {
@@ -74,28 +96,28 @@ const removeFromliked = async (id,dispatchAction,navigate) => {
           payload: response.data.likes,
         });
         Toast({
-        message: "Video removed from your liked videos.",
-        type: "success",
-      });
-      }
-      else{
-       
-          navigate("/signup");
-           Toast({
-        message: "Please login to continue.",
-        type: "warning",
-      });
-         
+          message: "Video removed from your liked videos.",
+          type: "success",
+        });
+      } else {
+        Toast({
+          message: "Some error occured, please try again later",
+          type: "error",
+        });
       }
     } catch (err) {
-      navigate("/signup");
       Toast({
-        message: "Please login to continue.",
-        type: "warning",
+        message: "Some error occured, please try again later",
+        type: "error",
       });
     }
-  };
+  } else {
+    navigate("/signup");
+    Toast({
+      message: "Please login to continue.",
+      type: "warning",
+    });
+  }
+};
 
-
-
-  export {isLiked, addToLiked,getLikedVideos,removeFromliked,}
+export { isLiked, addToLiked, getLikedVideos, removeFromliked };
