+++ src/contexts/video-action-context/video-action-utils/watchlater-utils.js
@@ -1,11 +1,12 @@
 import axios from "axios";
-import { Toast } from "../../../components/index"
+import { Toast } from "../../../components/index";
 
-const isInWatchLater = (id,data) => {
-  return data?.find(item=>item._id === id)
-}
+const isInWatchLater = (id, data) => {
+  return data?.find(item => item._id === id);
+};
 
-const getWatchLaterVideos = async (dispatchAction,navigate) => {
+const getWatchLaterVideos = async (isLoggedIn, dispatchAction, navigate) => {
+  if (isLoggedIn) {
     try {
       const response = await axios.get("/api/user/watchlater", {
         headers: {
@@ -13,24 +14,37 @@ const getWatchLaterVideos = async (dispatchAction,navigate) => {
         },
       });
       if (response.status === 200) {
-        dispatchAction({type:"UPDATE_WATCHLATER_VIDEOS", payload:response.data.watchlater})
+        dispatchAction({
+          type: "UPDATE_WATCHLATER_VIDEOS",
+          payload: response.data.watchlater,
+        });
       } else {
-        navigate("/signup");
+        Toast({
+          message: "Some error occured, please try again later",
+          type: "error",
+        });
       }
     } catch (err) {
-      navigate("/signup")
       Toast({
-        message: "Please login to continue.",
-        type: "warning",
+        message: "Some error occured, please try again later",
+        type: "error",
       });
     }
-  };
+  } else {
+    navigate("/signup");
+    Toast({
+      message: "Please login to continue.",
+      type: "warning",
+    });
+  }
+};
 
-const addToWatchLater = async (video,dispatchAction,navigate) => {
+const addToWatchLater = async (isLoggedIn, video, dispatchAction, navigate) => {
+  if (isLoggedIn) {
     try {
       const response = await axios.post(
         "/api/user/watchlater",
-        { video  },
+        { video },
         {
           headers: {
             authorization: localStorage.getItem("userToken"),
@@ -38,29 +52,42 @@ const addToWatchLater = async (video,dispatchAction,navigate) => {
         }
       );
       if (response.status === 201) {
-        dispatchAction({ type: "UPDATE_WATCHLATER_VIDEOS", payload: response.data.watchlater });
+        dispatchAction({
+          type: "UPDATE_WATCHLATER_VIDEOS",
+          payload: response.data.watchlater,
+        });
         Toast({
-        message: "Video added to watch later.",
-        type: "success",
-      });
-      }
-      else{
-          navigate("/signup")
-          Toast({
-        message: "Please login to continue.",
-        type: "warning",
-      });
+          message: "Video added to watch later.",
+          type: "success",
+        });
+      } else {
+        Toast({
+          message: "Some error occured, please try again later",
+          type: "error",
+        });
       }
     } catch (err) {
-     navigate("/signup")
-     Toast({
-        message: "Please login to continue.",
-        type: "warning",
+      Toast({
+        message: "Some error occured, please try again later",
+        type: "error",
       });
     }
-  };
+  } else {
+    navigate("/signup");
+    Toast({
+      message: "Please login to continue.",
+      type: "warning",
+    });
+  }
+};
 
-const removeFromWatchLater = async (id,dispatchAction,navigate) => {
+const removeFromWatchLater = async (
+  isLoggedIn,
+  id,
+  dispatchAction,
+  navigate
+) => {
+  if (isLoggedIn) {
     const path = `/api/user/watchlater/${id}`;
     try {
       const response = await axios.delete(path, {
@@ -74,26 +101,33 @@ const removeFromWatchLater = async (id,dispatchAction,navigate) => {
           payload: response.data.watchlater,
         });
         Toast({
-        message: "Video removed from watch later.",
-        type: "success",
-      });
-      }
-      else{
-       
-          navigate("/signup");
-           Toast({
-        message: "Please login to continue.",
-        type: "warning",
-      });
-         
+          message: "Video removed from watch later.",
+          type: "success",
+        });
+      } else {
+        Toast({
+          message: "Some error occured, please try again later",
+          type: "error",
+        });
       }
     } catch (err) {
-      navigate("/signup");
       Toast({
-        message: "Please login to continue.",
-        type: "warning",
+        message: "Some error occured, please try again later",
+        type: "error",
       });
     }
-  };
+  } else {
+    navigate("/signup");
+    Toast({
+      message: "Please login to continue.",
+      type: "warning",
+    });
+  }
+};
 
-  export {isInWatchLater, addToWatchLater,getWatchLaterVideos,removeFromWatchLater,}
+export {
+  isInWatchLater,
+  addToWatchLater,
+  getWatchLaterVideos,
+  removeFromWatchLater,
+};
